Fixed definition of type in var_name_size.

+++ src/sdf/data_file.cc
@@ -35,7 +35,7 @@ const enum DataFile::endianness DataFile::endian = BIG_ENDIAN_FILE;
 #else
 const enum DataFile::endianness DataFile::endian = LITTLE_ENDIAN_FILE;
 #endif
-const size_t DataFile::var_name_size = 64;
+const unsigned int DataFile::var_name_size = 64;
 
 /* Try to get lock. Return its file descriptor or -1 if failed.
  */
